@@ -1,9 +1,9 @@
 package talabaty.swever.com.online.Contact;
 
 import android.content.Context;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -69,7 +69,7 @@ public class ContactAdapter extends RecyclerView.Adapter<ContactAdapter.VHolder>
             holder.sell_image.setVisibility(View.INVISIBLE);
         }
         if (!products.get(position).getImage_url().isEmpty()) {
-            Picasso.with(mContext).load(products.get(position).getImage_url()).into(holder.image);
+            Picasso.get().load(products.get(position).getImage_url()).into(holder.image);
         }
 
         holder.action.setOnClickListener(new View.OnClickListener() {
