@@ -1,17 +1,16 @@
 package talabaty.swever.com.online.OffersDetails;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import talabaty.swever.com.online.Chart.Models.OperOrderModel;
+import talabaty.swever.com.online.Cart.Models.OperOrderModel;
 import talabaty.swever.com.online.R;
 
 public class OfferDetailsAdapter extends RecyclerView.Adapter<OfferDetailsAdapter.Vholder> {
