@@ -1,24 +1,20 @@
 package talabaty.swever.com.online.PrepareFood;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.CheckBox;
-import android.widget.CompoundButton;
-import android.widget.ImageView;
 import android.widget.TextView;
-
-import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
 import talabaty.swever.com.online.R;
-import talabaty.swever.com.online.Chart.*;
+import talabaty.swever.com.online.Cart.*;
 public class AdditionsAdapter extends RecyclerView.Adapter<AdditionsAdapter.Vholder> {
 
     Context context;
