@@ -2,14 +2,13 @@ package talabaty.swever.com.online.PrepareFood;
 
 import android.app.ProgressDialog;
 import android.content.Context;
-import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -43,8 +42,8 @@ import java.util.Map;
 
 import jp.wasabeef.recyclerview.adapters.SlideInBottomAnimationAdapter;
 import jp.wasabeef.recyclerview.animators.FadeInDownAnimator;
-import talabaty.swever.com.online.Chart.AdditionalModel;
-import talabaty.swever.com.online.Chart.Sanf;
+import talabaty.swever.com.online.Cart.AdditionalModel;
+import talabaty.swever.com.online.Cart.Sanf;
 import talabaty.swever.com.online.R;
 
 public class FragmentAdditions extends Fragment {
