@@ -1,20 +1,19 @@
 package talabaty.swever.com.online.ProductDetails;
 
 import android.content.Context;
-import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
 
 import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-import talabaty.swever.com.online.Chart.Sanf;
+import talabaty.swever.com.online.Cart.Sanf;
 import talabaty.swever.com.online.R;
 
 public class ProductDetailsAdapter extends RecyclerView.Adapter<ProductDetailsAdapter.Vholder> {
@@ -38,7 +37,7 @@ public class ProductDetailsAdapter extends RecyclerView.Adapter<ProductDetailsAd
     public void onBindViewHolder(@NonNull Vholder holder, final int position) {
 
         if (!talabats.get(position).getImage().isEmpty()) {
-            Picasso.with(context).load(talabats.get(position).getImage()).into(holder.image);
+           Picasso.get().load(talabats.get(position).getImage()).into(holder.image);
         }
     }
 
