@@ -2,10 +2,10 @@ package talabaty.swever.com.online.SubCategory;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -69,7 +69,7 @@ public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.Vholder>
             holder.sell_image.setVisibility(View.INVISIBLE);
         }
         if (!product.get(position).getImage_url().isEmpty()) {
-            Picasso.with(mContext).load(product.get(position).getImage_url()).into(holder.image);
+            Picasso.get().load(product.get(position).getImage_url()).into(holder.image);
         }
 
         holder.action.setOnClickListener(new View.OnClickListener() {
